wip: Integration tests using nunit
Here's the plan:

Separate integration tests from normal unit tests so that they're easier
to run separately, as well as easier to run via github workflow.

Challenges:

I can use MockFileSystem in the integration tests to remove dependency
on the real filesystem for accessing things like cache, configs, etc.
But the one thing I can't mock is the git command app. It's going to
want access to the real filesystem.

One option I considered is mocking the `IGitRepository` and other
classes. This would be a lot of work and potentially remove some of the
usefulness of the tests, but I should probably go this route when I
revisit.